refactor: User folder rename
